getting the cmdclass from versioneer, adding the custom clean command and then passing it to setup()
